# packages/openapi/src/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
paths:
...
  /users:
    post:
      tags:
        - User Resources
      summary: Register a new user
      description: Register a new user with the provided information
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithJWT'
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/NotAuthenticated"
        "409":
          $ref: "#/components/responses/AlreadyExists"
        "403":
          $ref: "#/components/responses/MissingPermissions"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: number
          example: 12928
          readOnly: true
        name:
          type: string
          example: John Doe
          maxLength: 50
        email:
          type: string
          format: email
      required:
        - id
        - name
        - email
    UserRegistration:
      allOf:
        - $ref: "#/components/schemas/User"
      properties:
        password:
          type: string
          example: password123
          minLength: 8
      required:
        - password
...
    UserWithJWT:
      allOf:
        - $ref: "#/components/schemas/User"
      properties:
        token:
          type: string
          description: JWT token for the user to use for sending requests
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.djwidj92e2.djwidj92e2
          readOnly: true
        tokenType:
          type: string
          description: The type of the jwt token
          example: Bearer
          readOnly: true
      required:
        - token
        - tokenType
...
    ServiceError:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong
        code:
          type: string
          $ref: '#/components/schemas/ErrorCodes'
          example: INTERNAL_SERVER_ERROR
        status:
          type: integer
          example: 500
      required:
        - message
        - code
        - status
    ErrorCodes:
      type: string
      enum:
        - INTERNAL_SERVER_ERROR
        - BAD_REQUEST
        - UNAUTHORIZED
        - NOT_FOUND
        - FORBIDDEN
        - CONFLICT
        - UNPROCESSABLE_ENTITY
        - TOO_MANY_REQUESTS
        - SERVICE_UNAVAILABLE
        - NETWORK_ERROR
        - CANCEL
        - UNKNOWN
        - INVALID_ARGUMENT
        - NOT_FOUND
        - ALREADY_EXISTS
        - PERMISSION_DENIED
        - FAILED_PRECONDITION
        - UNIMPLEMENTED
 # Responses - WebService Error Exceptions
  responses:
    NotFound:
      description: Resource has not been found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceError"
    AlreadyExists:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceError"
    NotAuthenticated:
      description: User is not authenticated.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceError"
    MissingPermissions:
      description: User lacks necessary permissions to access this resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceError"
    InvalidRequest:
      description: Body or Query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceError"
    NotImplemented:
      description: Routes or methods are not implemented.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceError"
...
    InternalServerError:
      description: Service returned Internal Server Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceError"